chore: bump op-succinct-elfs to v3.2.2-agglayer (#305)
This pull request updates a dependency version in the `Cargo.toml` file.
The `op-succinct-elfs` dependency is upgraded from tag `v3.1.0-agglayer`
to `v3.2.2-agglayer`, ensuring the project uses the latest features and
fixes from that library.

- Dependency update:
  * Upgraded `op-succinct-elfs` to tag `v3.2.2-agglayer` in `Cargo.toml`

BREAKING-CHANGE: Short description of the breaking changes.
This PR will update the elfs therefore the vkeys

---------

Co-authored-by: atanmarko <atanmarko@users.noreply.github.com>